Refactored CBPE into single class
